Fix nondeterministic pipe failure.
When a pipe is opened in write-only nonblocking mode,
it fails if the other end hasn't been opened for reading yet.

This happened occasionally because the client was
sending the command that includes the pipe's path
before opening the pipe for reading.

The simplest way to avoid this is to always open the read pipe
before sending the command.