Refactor the secure attestation socket in Pelz to fix sgx errors.
Previously, the trusted request handler would fail due to a "nested ecall" issue
because we were using the pattern ecall->ocall->ecall, which is not allowed.
With the new secure attestation protocol,
the request message is initially decrypted within the enclave,
so it would be natural to handle it in trusted code,
but there are issues with this approach:
1. Third-party libraries like cjson and uriparser can't be used in trusted code,
   so they have to be wrapped in ocalls, which compromises the enclave data protections.
2. It's difficult to use existing untrusted code from the new trusted handler.
   Wrapping in an ocall can create "nested ecall" issues,
   which must then be fixed by converting the untrusted code to trusted code
   (or duplicating it if it's already being used in an untrusted context).

Now, the request handler for secure attested messages is primarily untrusted,
which allows for much more flexibility in using existing untrusted code.